requirements.txt
numpy==1.23.5
ultralytics>=8.0.0
supervision>=0.18.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.36.0
faiss-cpu>=1.7.4
Pillow>=10.0.0
scikit-learn>=1.3.0 